change pin assignments, move to polling model
This commit changes the PIN assignments to the new setup. More
importantly, it moves to a polling model. Previously, somebody could
flood the server with alternating requests and make the hardware
rapidly oscillate, which sounds like a terrible idea to me and might
allow someone to break hardware. Now, requests only update a variable
which gets checked for changes every minute.