use interrupts to detect LED changes
This commit moves from a polling model (reading the GPIO pins whenever a
request is made) to an event driven model (reacting to the GPIO pins
changing). To do this, it now spawns a thread and uses `IO::select` to
block until the pins have changed, then updates a variable. Now all
requests are serviced purely from variables in memory instead of reading
files, which should be much more performant since the LED's shouldn't be
updating as often as requests are being made.

This work also lays the foundation for the eventual message queuing.